chore: setup a basic test with vitest and organise types directory
