+++ ui/components/chat/tests/FileDisplay.test.tsx
@@ -1,0 +1,8 @@
+
+import { describe, it, expect } from 'vitest'
+
+describe('FileDisplay', () => {
+  it('should run a test', () => {
+    expect(true).toBe(true)
+  })
+})
