+++ ui/tsconfig.json
@@ -23,9 +23,7 @@
       }
     ],
     "paths": {
-      "@/*": [
-        "./*"
-      ]
+      "@/*": ["./*"],
     }
   },
   "include": [
