+++ ui/types/FileAttachment.ts
@@ -1,0 +1,9 @@
+interface FileAttachment {
+  type: 'file'
+  mediaType: string
+  filename: string
+  url: string
+  size?: number
+}
+
+export type { FileAttachment };
