+++ ui/types/FlowInputValue.ts
@@ -1,0 +1,3 @@
+type FlowInputValue = string | number | boolean | null | undefined
+
+export type { FlowInputValue };
