+++ ui/types/FlowInputValues.ts
@@ -1,0 +1,5 @@
+import type { FlowInputValue } from ".";
+
+type FlowInputValues = Record<string, FlowInputValue>
+
+export type { FlowInputValue, FlowInputValues };
