+++ ui/types/OpenAPISpec.ts
@@ -1,0 +1,5 @@
+import { OpenAPIV3_1 } from "openapi-types";
+
+type OpenAPISpec = OpenAPIV3_1.Document
+
+export type { OpenAPISpec };
