+++ ui/types/ResponseData.ts
@@ -1,0 +1,3 @@
+type ResponseData = Record<string, unknown>
+
+export type { ResponseData };
