+++ ui/types/SchemaProperty.ts
@@ -1,0 +1,7 @@
+import { OpenAPIV3_1 } from "openapi-types";
+
+type SchemaProperty = OpenAPIV3_1.SchemaObject & {
+  qtype_type?: string
+}
+
+export type { SchemaProperty };
